requirements.txt
lxml
PyOpenSSL
aiodns
